+++ Sora/TLSSecurityPolicy.swift
@@ -1,11 +1,11 @@
 import Foundation
 import WebRTC
 
-private var tlsSecurityPolicyTable: [TLSSecurityPolicy: RTCTlsCertPolicy] =
+private let tlsSecurityPolicyTable: [TLSSecurityPolicy: RTCTlsCertPolicy] =
   [.secure: .secure, .insecure: .insecureNoCheck]
 
 /// TLS のセキュリティポリシーを表します。
-public enum TLSSecurityPolicy {
+public enum TLSSecurityPolicy: Sendable {
   /// サーバー証明書を確認します。
   case secure